chore: update shelljs (#21)
This updates ShellJS from "^0.7.0" to "^0.8.4". This is necessary to
resolve issue #20 (the node 14 circular dependency warning). Off the top
of my head, I can't remember what breaking changes landed in v0.8.*, so
I'm hoping this should be safe (it seems to pass automated tests).

This also updates appveyor and travis to test the node LTS releases
(even-numbered releases) from 6 through 14.